% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.deconv-package.R
\docType{package}
\name{isr.deconv-package}
\alias{isr.deconv-package}
\title{isr.deconv: Insulin Secretion Rate Deconvolution}
\description{
Calculates insulin secretioin rates from C-peptide values based on the methods described in Van Cauter et al. (1992), "Estimation of Insulin Secretion Rates from C-Peptide Levels". Included functions calculate estimated insulin secretion rates using linear or cubic spline interpolation of c-peptide values (see Eaton et al., 1980, and Polonsky et al., 1986) and to calculate estimates of input coefficients (volume of distribution, short half life, long half life, and fraction attributed to short half life) as described by Van Cauter.
}
\author{
\strong{Maintainer}: Ken Stier \email{kennethstier99@gmail.com} [copyright holder]

}
\keyword{internal}
