% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.fraction.R
\name{isr.fraction}
\alias{isr.fraction}
\title{C-Peptide Fraction Attributable to Short Half Life Lookup}
\usage{
isr.fraction(subject.type = c("normal", "obese", "niddm"))
}
\arguments{
\item{subject.type}{String for patient type.}
}
\value{
Numeric value.
}
\description{
Outputs the relevant fraction attributable to short half life for use in
\link{isr.deconv} based on patient type (normal, obese, or non-insulin dependent
diabetes mellitus).
}
\examples{
isr.fraction("normal")

isr.fraction("obese")

isr.fraction("niddm")

}
\seealso{
\code{\link[=isr.deconv]{isr.deconv()}}
}
