% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.longHL.R
\name{isr.longHL}
\alias{isr.longHL}
\title{C-Peptide Long Half Life Calculator}
\usage{
isr.longHL(subject.age)
}
\arguments{
\item{subject.age}{Numeric for patient age in years.}
}
\value{
Numeric value.
}
\description{
Outputs the long half life based on patient age for use in \code{\link[=isr.deconv]{isr.deconv()}} per
the Van Cauter method of estimating insulin secretion rate.
}
\examples{
isr.longHL(18.08)

}
\seealso{
\code{\link[=isr.deconv]{isr.deconv()}}
}
