% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isr.shortHL.R
\name{isr.shortHL}
\alias{isr.shortHL}
\title{C-Peptide Short Half Life Calculator}
\usage{
isr.shortHL(subject.type = c("normal", "obese", "niddm"))
}
\arguments{
\item{subject.type}{String for patient type.}
}
\value{
Numeric value.
}
\description{
Outputs the relevant long half life of c-peptide for use in \code{\link[=isr.deconv]{isr.deconv()}}
based on patient type (normal, obese, or non-insulin dependent diabetes
mellitus).
}
\examples{
isr.shortHL("normal")

isr.shortHL("obese")

isr.shortHL("niddm")

}
\seealso{
\code{\link[=isr.deconv]{isr.deconv()}}
}
